Implementhelper functions for future k moves
